requirements.txt
django==4.1.7
djangorestframework==3.14.0
dj-rest-auth==3.0.0
django-allauth==0.53.1